requirements.txt
streamlit
opencv-python-headless
mediapipe==0.10.9
numpy
fpdf2